One Front Door/SourceCode/DPSG.Web.Merchandiser/src/WebApp/src/app/merc-main/merc-main.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { MercMainComponent } from './merc-main.component';

describe('Component: MercMain', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [MercMainComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([MercMainComponent],
      (component: MercMainComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(MercMainComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(MercMainComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-merc-main></app-merc-main>
  `,
  directives: [MercMainComponent]
})
class MercMainComponentTestController {
}

